feat #32: starCount 사용하는 값 stars 값으로 모두 변경 후, 리뷰 평균값 형을 float -> double 수정함

+++ core/src/main/java/com/eoyeongbooyeong/core/designsystem/component/star/MyReview.kt
@@ -31,7 +31,7 @@ fun MyReview(
     reviewId: Int = -1,
     writerId: Int = -1,
     storeName: String = "",
-    reviewScore: Float = 0f,
+    reviewScore: Double = 0.0,
     reviewContent: String = "",
     reviewDate: String = "",
 ) {
@@ -90,7 +90,7 @@ fun MyReviewPreview() {
     BooTheme {
         MyReview(
             storeName = "아홉산 숲",
-            reviewScore = 4.5f,
+            reviewScore = 4.5,
             reviewContent = "생각보다 내부가 엄청 넓었어요. 천천히 산책하면서 사진찍으려면 1시간30분~2시간은 잡아야 가능할 곳이에요.",
             reviewDate = "2021.10.10",
             modifier = Modifier.background(White),
