+++ domain/src/main/java/com/eoyeongbooyeong/domain/entity/ReviewInfoEntity.kt
@@ -4,7 +4,7 @@ data class ReviewInfoEntity(
     val id: Int = -1,
     val placeId: Int = -1,
     val reviewContent: String = "",
-    val reviewScore: Float = 0f,
+    val reviewScore: Double = 0.0,
     val writerId: Int = -1,
     val writerNickName: String = "",
     val createdAt: String = "",
